requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blinker==1.6.3
Brlapi==0.8.3
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
contourpy==1.1.0
crispy-bootstrap5==0.7
crypto==1.4.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
Django==4.2.2
django-ajax-selects==2.2.0
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-markdownify==0.9.3
django-markdownx==4.0.2
django-silk==5.0.4
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.2
eyeD3==0.8.10
fasteners==0.18
filelock==3.12.2
fire==0.5.0
Flask==3.0.0
fonttools==4.40.0
frozenlist==1.3.3
future==0.18.3
gprof2dot==2022.7.29
grpcio==1.30.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
identify==2.5.24
idna==3.4
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
markdown-it-py==3.0.0
markdownify==0.11.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==8.10.0
MouseInfo==0.1.3
multidict==6.0.4
mutagen==1.45.1
mypy-extensions==1.0.0
Naked==0.1.32
nemo-emblems==5.8.0
netaddr==0.8.0
netifaces==0.11.0
nodeenv==1.8.0
numpy==1.25.0
oauthlib==3.2.0
onboard==1.4.1
openai==0.27.8
opencv-python==4.8.0.74
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
PAM==0.4.2
patch-ng==1.17.4
pathspec==0.11.1
pexpect==4.8.0
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.7.0
pluggy==1.2.0
pre-commit==3.3.3
prettytable==3.9.0
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
PyAutoGUI==0.9.53
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
pygame==2.5.0
PyGetWindow==0.0.9
Pygments==2.15.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.0.9
pyparted==3.11.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.3.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
python3-xlib==0.15
pytweening==1.0.4
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0
pyzk==0.9
pyzkaccess==1.0
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
rich==13.4.2
ruff==0.0.272
scour==0.38.2
SecretStorage==3.3.1
selenium==4.10.0
setproctitle==1.2.2
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
systemd-python==234
termcolor==2.3.0
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
tomli==2.0.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.16
vboxapi==1.0
virtualenv==20.23.1
wadllib==1.3.6
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
Werkzeug==3.0.0
wrapt==1.15.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarg==0.1.9
yarl==1.9.2
yt-dlp==2022.4.8
zipp==1.0.0
